fix(exporter): repair empty byline at excerpt source (#800)
