+++ crates/webfang_core/src/application/extraction.rs
@@ -302,7 +302,9 @@ pub async fn extract_content(
                 title: crate::application::resolve_title(&article.title, url),
                 content: article.text_content,
                 url: ValidUrl::new(url.clone()),
-                excerpt: article.excerpt,
+                excerpt: article.excerpt.as_deref().map(|e| {
+                    crate::domain::excerpt_repair::repair_empty_byline(e, author.as_deref())
+                }),
                 author,
                 date: article.published_time,
                 // Store CLEAN HTML from Readability (not raw HTML with nav/ads/footer)
