+++ crates/webfang_core/src/domain/mod.rs
@@ -44,6 +44,8 @@ pub mod value_objects;
 
 pub mod content_processor;
 pub mod embedding_port;
+/// Shared excerpt byline-repair invariant (regex + whitespace), pure & IO-free.
+pub(crate) mod excerpt_repair;
 pub mod note_repository;
 pub mod semantic_cleaner;
 pub mod semantic_inspector;
